# openapi/schema.openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "0.0.1",
        "title": "Iglu API",
        "description": "Iglu API Docs for use in your development"
    },
    "servers": [{ "url": "/" }],
...
    "paths": {
...
        "/api/v1/deploy/deployment/{deployment_id}": {
            "get": {
                "description": "Get an overview of a given deployment (by deployment id)",
                "summary": "Get deployment overview",
                "tags": ["api/v1/deploy", "cachix"],
                "parameters": [
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "deployment_id",
                        "in": "path"
                    },
                    {
                        "schema": { "type": "string" },
                        "required": true,
                        "name": "authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Information about the deployment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "closureSize": { "type": "number", "nullable": true },
                                        "createdOn": { "type": "string", "format": "date-time" },
                                        "finishedOn": {
                                            "type": "string",
                                            "nullable": true,
                                            "format": "date-time"
                                        },
                                        "id": { "type": "string", "format": "uuid" },
                                        "index": { "type": "number" },
                                        "startedOn": { "type": "string", "format": "date-time" },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "InProgress",
                                                "Pending",
                                                "Cancelled",
                                                "Failed",
                                                "Succeeded"
                                            ]
                                        },
                                        "storePath": { "type": "string" }
                                    },
                                    "required": [
                                        "closureSize",
                                        "createdOn",
                                        "finishedOn",
                                        "id",
                                        "index",
                                        "startedOn",
                                        "status",
                                        "storePath"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authorized to view deployment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The deployment does not exist on the server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status_code": { "type": "number" },
                                        "status_message": { "type": "string" },
                                        "is_error": { "type": "boolean" },
                                        "data": {
                                            "type": "object",
                                            "properties": { "error_details": { "type": "string" } },
                                            "required": ["error_details"]
                                        },
                                        "timestamp": { "type": "number" }
                                    },
                                    "required": [
                                        "status_code",
                                        "status_message",
                                        "is_error",
                                        "data",
                                        "timestamp"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}